Added Plugin.saveDefaultConfig() method
Turning on options.copyDefaults(true) and then calling Plugin.saveConfig() is the existing method for saving the embedded config.yml
file to disk, but in the process all formatting, spacing, and comments are stripped from the file. The Plugin.SaveDefaultConfig() method
writes the contents of the default config.yml file to the default configuration location verbatim. The result is a default config.yml
file with all formatting and comments preserved.

* If the default configuration directory is absent, SaveDefaultConfig() will create it.
* If the default config.yml is missing from the plugin, SaveDefaultConfig() will write out a blank file.
* If a config.yml file already exists in the target directory, SaveDefaultConfig() will not overwrite it and will write a
warning in the log.